First Upload :+1:
Hello everyone,

Today I want to share a brand new project which I'm working on it since two weeks at least.
The crypto-world is fascinating , that's why I want to mix all the thing that I love together : Dev' + Trading = :heart_eyes: .

So this is the first upload and more will come in the future as the project moving on.

See you really soon,

Franck :bicyclist: .