The Graph Feature is here ! :grapes:
Hello dear Users,

Today I update my program to give a new feature.
This Feature allow you to see a Chart by typing a Coin/Token in a textbox.
You have to type the asset , the starting Date and the Ending Date and voilà , a Grapg will appear on your screen.

You have to know that more the period is big, more it's will may takes to render the chart for your computer ; it's depends of it's power.

Thank You :fireworks: ,
Franck :night_with_stars: .